refactor: move ProjectDetailComponent to secondary-components directory
