+++ src/app/secondary-components/project-detail/project-detail.component.html
